[HOTSPOT-181] fix: dataLimit dto에서 long으로 변경

+++ src/main/java/hotspot/user/family/controller/response/UpdateDataLimitResponse.java
@@ -13,7 +13,7 @@ import lombok.Builder;
 public record UpdateDataLimitResponse(
         Long familyId,
         Long subId,
-        double dataLimit,
+        long dataLimit,
         boolean isLocked
 ) {
 }
